[REFACTOR][TEST] Fold fallback round-trip tests into per-backend codegen tests
Replace the 7 tests/python/runtime/test_<backend>_fallback_module.py
files with one export+load test per backend blended into the
existing tests/python/codegen/test_target_codegen_<backend>.py.
Tests reuse the file's existing fixture / decorator / target-string
patterns. Hexagon and WebGPU intentionally have no fallback test
for now — their standalone files are deleted with no replacement.
