--- tests/python/runtime/test_hexagon_fallback_module.py
@@ -1,76 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-"""Tests for the per-backend Hexagon fallback module.
-
-Hexagon codegen requires LLVM with the Hexagon target AND the Hexagon
-toolchain.  Even with both available, exercising the codegen pipeline on
-a non-Hexagon host requires a Hexagon-shaped kernel (specific target
-attrs, kernel with thread bindings, etc.) — too far afield for a
-fallback-shape test.
-
-Instead this test reaches the Hexagon real-runtime FFI factory
-"ffi.Module.create.hexagon" directly: when USE_HEXAGON=ON the real
-``HexagonModuleNode`` is constructed; when USE_HEXAGON=OFF the registry
-key is missing on this side, but the same byte-format SaveToBytes/
-LoadFromBytes round-trip is verified by the C++ unit tests for the
-fallback class.
-
-Here, with USE_HEXAGON=OFF on this CI host, we can still verify:
-
-  * the fallback factory is reachable by name through tvm-ffi
-  * a constructed fallback module reports kind=="hexagon" and fmt=="so"
-  * GetFunction raises a clear "runtime not linked" error
-  * SaveToBytes produces the expected 3-field shape that a USE_HEXAGON=ON
-    receiver can deserialize (verification of the receiving side is left
-    for a USE_HEXAGON=ON CI run).
-"""
-
-import pytest
-
-import tvm
-import tvm.testing
-
-
-def _hexagon_runtime_loader_registered():
-    """The real runtime registers ffi.Module.load_from_bytes.hexagon when
-    USE_HEXAGON=ON.  Absence => USE_HEXAGON=OFF host (typical CI)."""
-    return tvm.get_global_func("ffi.Module.load_from_bytes.hexagon", allow_missing=True) is not None
-
-
-@pytest.mark.skipif(
-    _hexagon_runtime_loader_registered(),
-    reason="USE_HEXAGON=ON in this build — fallback-shape test is for OFF builds",
-)
-def test_hexagon_fallback_construction_and_save():
-    """Without the real runtime registered, the codegen-side wrapper falls
-    through to the fallback.  Verify the fallback path on a USE_HEXAGON=OFF
-    host by reaching it via the C++ side through ``ffi.Module.create.hexagon``
-    — but since the real loader is not registered, the codegen-side
-    wrapper would fall through to the fallback.  We can't directly call
-    the wrapper from Python (no FFI binding), so we approximate with a
-    minimal smoke test: confirm the registry name is *absent*, which is
-    the precondition for the fallback-on-miss path.
-    """
-    # Precondition: the create-key is absent (USE_HEXAGON=OFF), so any
-    # codegen-side caller would hit the fallback branch.
-    assert tvm.get_global_func("ffi.Module.create.hexagon", allow_missing=True) is None
-    # Precondition: the load-from-bytes key is absent for the same reason.
-    assert tvm.get_global_func("ffi.Module.load_from_bytes.hexagon", allow_missing=True) is None
-
-
-if __name__ == "__main__":
-    tvm.testing.main()
